+++ spec/lib/tmuxinator/project_spec.rb
@@ -28,6 +28,10 @@ describe Tmuxinator::Project do
     FactoryGirl.build(:nameless_window_project)
   end
 
+  it "should include Hooks" do
+    expect(project).to be_kind_of(Tmuxinator::Hooks::Project)
+  end
+
   describe "#initialize" do
     context "valid yaml" do
       it "creates an instance" do
@@ -166,7 +170,8 @@ describe Tmuxinator::Project do
         rendered = project_with_literals_as_window_name
         expect(rendered.windows.map(&:name)).to match_array(
           %w(222 222333 111222333444555666777 222.3 4e5 4E5
-             true false nil // /sample/))
+             true false nil // /sample/)
+        )
       end
     end
   end